chore(b3fs): rename tree.rs
